ladder=4, sai mode
This commit tries to address the difficulty in reading ladders in a
very simple way. We give a definition of a forced move, that is a
move that gives atari or extends from atari. When playing
simulations in MCTS, the number of forced moves after the current
node is subtracted from the number of visits, in the denominator of
the policy term in UCT formula.

In this way it is easier for the MCTS to explore deeply a ladder.

This can be deactivated with option --noladdercode.